feat: refactor all fetch & mutation with server actions
